Expand OWs to u16
This is based on ghoulslash's expansion branch, but he missed many places where u8s needed to be changed into u16s and didn't removed inConnection to make his work. This version should've caught everything and maintained the CD clone object/inConnection functionality.

+++ include/union_room.h
@@ -88,7 +88,7 @@ struct WirelessLink_Group
 struct UnionRoomObject
 {
     u8 state;
-    u8 gfxId;
+    u16 gfxId;
     s8 animState;
     u8 schedAnim;
 };
