+++ src/data/object_events/berry_tree_graphics_tables.h
@@ -418,9 +418,9 @@ static const struct SpriteFrameImage sPicTable_TamatoBerryTree[] = {
 
 const u8 gBerryTreePaletteSlotTable_Tamato[] = {3, 4, 2, 2, 2};
 
-const u8 gDeadBerryTreeObjectEventGraphicsIdTable[] = {OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES};
+const u16 gDeadBerryTreeObjectEventGraphicsIdTable[] = {OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES};
 
-const u8 gBerryTreeObjectEventGraphicsIdTable[] = {OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_LATE_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_LATE_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_LATE_STAGES};
+const u16 gBerryTreeObjectEventGraphicsIdTable[] = {OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_EARLY_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_LATE_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_LATE_STAGES, OBJ_EVENT_GFX_ZBERRY_TREE_LATE_STAGES};
 
 const struct SpriteFrameImage *const gBerryTreePicTablePointers[] = {
     [ITEM_CHERI_BERRY - FIRST_BERRY_INDEX]  = sPicTable_CheriBerryTree,
@@ -514,7 +514,7 @@ const u8 *const gBerryTreePaletteSlotTablePointers[] = {
     [ITEM_ENIGMA_BERRY - FIRST_BERRY_INDEX] = gBerryTreePaletteSlotTable_Durin,
 };
 
-const u8 *const gBerryTreeObjectEventGraphicsIdTablePointers[] = {
+const u16 *const gBerryTreeObjectEventGraphicsIdTablePointers[] = {
     [ITEM_CHERI_BERRY - FIRST_BERRY_INDEX] = gBerryTreeObjectEventGraphicsIdTable,
     [ITEM_CHESTO_BERRY - FIRST_BERRY_INDEX] = gBerryTreeObjectEventGraphicsIdTable,
     [ITEM_PECHA_BERRY - FIRST_BERRY_INDEX] = gBerryTreeObjectEventGraphicsIdTable,
